Crate spl_token_confidential_transfer_proof_generation

Modules

burn

Generates the zero-knowledge proofs required for a confidential burn.

encryption

errors

mint

Generates the zero-knowledge proofs required for a confidential mint.

transfer

Generates the zero-knowledge proofs required for a confidential transfer.

transfer_with_fee

Generates the zero-knowledge proofs required for a confidential transfer with a fee.

withdraw

Generates the zero-knowledge proofs required for a confidential withdraw.

Structs

CiphertextValidityProofWithAuditorCiphertext

A type that wraps a ciphertext validity proof along with two lo and hi ciphertexts.

Constants

REMAINING_BALANCE_BIT_LENGTH

The bit length of the encrypted remaining balance in a token account

TRANSFER_AMOUNT_HI_BITS

The high bit length of the encrypted transfer amount

TRANSFER_AMOUNT_LO_BITS

The low bit length of the encrypted transfer amount

Functions

try_combine_lo_hi_ciphertexts

try_combine_lo_hi_commitments

try_combine_lo_hi_openings

try_combine_lo_hi_u64Deprecated

Combine two numbers that are interpreted as the low and high bits of a target number. The bit_length parameter specifies the number of bits that amount_hi is to be shifted by.

try_split_u64

Takes in a 64-bit number amount and a bit length bit_length. It returns: